+++ smart-home-planner/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - More device fields: serial number, purchase date, store, price and Warranty Expiration
+- Added Global Search
 
 ## 1.0.0 - 2026-02-21